Improve docs-specs setup
Now that we have a first template, remove the example spec that was only
added so that we had one specification and thus could gate.

Improve the index page to include the README and reference both
sample template and empty skeleton.

Enhance README to include URLs about this repo and where it gets
published, also make URLs clickable - and not verbatim text.

Change-Id: I1627db3cfd6bdd44e1dd35ad1c18ec8db633056e