Move driver docs out of openstack-manuals
As discussed via email and at the OpenStack Kilo summit, move most of
the driver docs out of the tree and only add links to them.

The in-tree content will be a small standard format with links to
external pages - and include the autogenerated configure options.

The goal of this spec is to get consistent and version independend
driver documentation. It will give driver authors more flexibility in
changing the documentation while removing work for both documentation
and driver teams.

Change-Id: I80f420e77d0f6036800af4099d274149e23971e0