Icehouse release for Training Guides
As discussed in the openstack summit, to move closer to our goals for
delivering openstack training, we concluded to move away from XML book
format and adopt RST slides for the same.

The goal of this spec is to provide description and required steps to
carry out the actions. This spec explains why Icehouse release is
important for training guides.

Change-Id: Icdbcf5f42da6c0aff3548f6d13109ebb3bbda7b7