Create seperate OpenStack Labs repository
OpenStack Labs is the training-labs under training-guides. Since the
Icehouse release, labs section is moving faster than the rest of the
training guides and does not really sync with the documentation content
in there. Thus, giving the scripts a new home will make it easier
to develop, test and release the scripts.

Change-Id: I15856f09d582d8c298eafa80e106f3a70be3743b