Create training-labs repository
Training labs is the labs folder under training-guides. Since the
Icehouse release, labs section is moving faster than the rest of the
training guides and does not really sync with the documentation content
in there. Thus, giving the scripts a new home will make it easier
to develop, test and release the scripts.

This patch is to address the change of name from openstack-labs.
  * Deletes openstacklabs spec.

Change-Id: Ibddf1191acb7d9cf81f1321699cc9b0039ddbb1e