Project Specific Install Guides
This is a proposal that allows project teams to write their own Install
Guides and properly promote them. The goal is to have a proposal that
does not cause significant load on the documentation team and gives main
responsibility to the project team. Instead the doc team should act as
enabler.

Change-Id: Ic028303dd42f055d284043ac5bec3854ef6769de